requirements.txt
grpcio==1.64.1
protobuf==5.27.3
